Pull up `LocationPath`
